+++ mock/service.go
@@ -9,7 +9,7 @@ type APIClient struct {
 	GetSelfFn func(token string) (*instagram.UserResponse, error)
 	GetSelfInvoked bool
 
-	GetRecentMediaFn      func(token string, maxID string, minID string, count int64) (*instagram.RecentMediaResponse, error)
+	GetRecentMediaFn      func(token string, maxID string, minID string, count int) (*instagram.RecentMediaResponse, error)
 	GetRecentMediaInvoked bool
 }
 
@@ -21,7 +21,7 @@ func NewAPIClient() *APIClient {
 		GetSelfFn: func(token string) (*instagram.UserResponse, error) {
 			return &instagram.UserResponse{}, nil
 		},
-		GetRecentMediaFn: func(token string, maxID string, minID string, count int64) (*instagram.RecentMediaResponse, error) {
+		GetRecentMediaFn: func(token string, maxID string, minID string, count int) (*instagram.RecentMediaResponse, error) {
 			return &instagram.RecentMediaResponse{}, nil
 		},
 	}
@@ -37,7 +37,7 @@ func (c *APIClient) GetSelf(token string) (*instagram.UserResponse, error) {
 	return c.GetSelfFn(token)
 }
 
-func (c *APIClient) GetRecentMedia(token string, maxID string, minID string, count int64) (*instagram.RecentMediaResponse, error) {
+func (c *APIClient) GetRecentMedia(token string, maxID string, minID string, count int) (*instagram.RecentMediaResponse, error) {
 	c.GetRecentMediaInvoked = true
 	return c.GetRecentMediaFn(token, maxID, minID, count)
 }
